refactor(paint): rename to PaintTool
